+++ cypress/elements/optionsModal/legend.js
@@ -3,8 +3,9 @@ const legendKeyOptionEl = 'option-legend-key'
 const legendKeyEl = 'visualization-legend-key'
 const legendKeyContainerEl = 'legend-key-container'
 const legendKeyItemEl = 'legend-key-item'
-const visualizationContainerEl = 'visualization-container'
-const singleValueOutputEl = 'visualization-primary-value'
+const singleValueTextEl = 'visualization-primary-value'
+const singleValueIconEl = 'visualization-icon'
+const singleValueOutputEl = 'visualization-container'
 const legendDisplayStrategyByDataItemEl = 'legend-display-strategy-BY_DATA_ITEM'
 const legendDisplayStrategyFixedEl = 'legend-display-strategy-FIXED'
 const legendDisplayStyleOptionTextEl = 'legend-display-style-option-TEXT'
@@ -24,10 +25,22 @@ export const changeDisplayStrategyToFixed = () =>
         .contains('Select a single legend for the entire visualization')
         .click()
 
+export const changeDisplayStrategyToByDataItem = () =>
+    cy
+        .getBySel(optionsModalContentEl)
+        .contains('Use pre-defined legend per data item')
+        .click()
+
 export const changeDisplayStyleToText = () =>
     cy
         .getBySel(optionsModalContentEl)
         .contains('Legend changes text color')
+        .click()
+
+export const changeDisplayStyleToFill = () =>
+    cy
+        .getBySel(optionsModalContentEl)
+        .contains('Legend changes background color')
         .click()
 
 export const expectLegendToBeEnabled = () =>
@@ -65,20 +78,23 @@ export const changeFixedLegendSet = (legendSetName) => {
 export const expectFixedLegendSetToBe = (legendSetName) =>
     cy.getBySel(fixedLegendSetSelectEl).should('contain', legendSetName)
 
+export const expectSingleValueToHaveTextColor = (color) =>
+    cy.getBySel(singleValueTextEl).invoke('attr', 'fill').should('eq', color)
+
+export const expectSingleValueToNotHaveBackgroundColor = () =>
+    cy.getBySel(singleValueOutputEl).should('not.have.attr', 'style')
+
 export const expectSingleValueToHaveBackgroundColor = (color) =>
     cy
-        .getBySel(visualizationContainerEl)
-        .parent()
-        .should('have.css', 'background-color', color)
+        .getBySel(singleValueOutputEl)
+        .invoke('attr', 'style')
+        .should('contain', `background-color: ${color}`)
 
-export const expectSingleValueToNotHaveBackgroundColor = (color) =>
+export const expectSingleValueToHaveIconColor = (color) =>
     cy
-        .getBySel(visualizationContainerEl)
-        .parent()
-        .should('not.have.css', 'background-color', color)
-
-export const expectSingleValueToBeColor = (color) =>
-    cy.getBySel(singleValueOutputEl).invoke('attr', 'fill').should('eq', color)
+        .getBySel(singleValueIconEl)
+        .invoke('attr', 'style')
+        .should('contain', `color: ${color}`)
 
 export const toggleLegendKeyOption = () =>
     cy.getBySel(optionsModalContentEl).contains('Show legend key').click()